requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
streamlit
pandas
requests
python-dotenv
itsdangerous
httpx
openai